+++ test/which.test.js
@@ -11,7 +11,7 @@ describe('which()', () => {
 
   it('should return the path of the `executable.cmd` file on Windows', async () => {
     try {
-      let executable = await which('executable', false, options);
+      let executable = await which('executable', Object.assign({all: false}, options));
       if (Finder.isWindows) expect(executable.endsWith('\\test\\fixtures\\executable.cmd')).to.be.true;
       else expect(true).to.not.be.ok;
     }
@@ -24,7 +24,7 @@ describe('which()', () => {
 
   it('should return all the paths of the `executable.cmd` file on Windows', async () => {
     try {
-      let executables = await which('executable', true, options);
+      let executables = await which('executable', Object.assign({all: true}, options));
       if (!Finder.isWindows) expect(true).to.not.be.ok;
       else {
         expect(executables).to.be.an('array').and.have.lengthOf(1);
@@ -40,7 +40,7 @@ describe('which()', () => {
 
   it('should return the path of the `executable.sh` file on POSIX', async () => {
     try {
-      let executable = await which('executable.sh', false, options);
+      let executable = await which('executable.sh', Object.assign({all: false}, options));
       if (Finder.isWindows) expect(true).to.not.be.ok;
       else expect(executable.endsWith('/test/fixtures/executable.sh')).to.be.true;
     }
@@ -53,7 +53,7 @@ describe('which()', () => {
 
   it('should return all the paths of the `executable.sh` file on POSIX', async () => {
     try {
-      let executables = await which('executable.sh', true, options);
+      let executables = await which('executable.sh', Object.assign({all: true}, options));
       if (Finder.isWindows) expect(true).to.not.be.ok;
       else {
         expect(executables).to.be.an('array').and.have.lengthOf(1);